Added support to the placed sappers, quality of life changes
Now if the sapper has a custom model it'll show up on a sapped sentry as well. Some quality of life changes ( previously I waited a frame whenever post_inventory_application was called, not a big deal but it's nicer now ), removed some stuff, added the base code for shields ( not supported for the time being ).